+++ comb_spec_searcher/specification.py
@@ -4,11 +4,13 @@ where each of the bi appear exactly once on the left hand side of some rule.
 """
 import logging
 from copy import copy
-from typing import Dict, Generic, Iterable, Iterator, Sequence, Tuple
+from functools import reduce
+from operator import mul
+from typing import Dict, Generic, Iterable, Iterator, List, Sequence, Tuple
 
 import sympy
 from logzero import logger
-from sympy import Eq, Expr, Function, solve, var
+from sympy import Eq, Expr, Function, Number, solve, var
 
 from .combinatorial_class import (
     CombinatorialClass,
@@ -35,7 +37,13 @@ from .strategies import (
     VerificationStrategy,
 )
 from .strategies.rule import AbstractRule
-from .utils import DisableLogging, RecursionLimit, maple_equations, taylor_expand
+from .utils import (
+    DisableLogging,
+    RecursionLimit,
+    maple_equations,
+    pretty_print_equations,
+    taylor_expand,
+)
 
 __all__ = ("CombinatorialSpecification",)
 
@@ -148,6 +156,7 @@ class CombinatorialSpecification(
                     AlreadyVerified(self.rules_dict), apply_first=True
                 ),
             )
+            logger.info(css.run_information())
             while True:
                 try:
                     css.do_level()
@@ -244,6 +253,30 @@ class CombinatorialSpecification(
                     sympy.Function("NOTIMPLEMENTED")(x),
                 )
 
+    def get_initial_conditions(self, check: int = 6) -> List[Expr]:
+        """
+        Compute the initial conditions of the root class. It will use the
+        `count_objects_of_size` method if implemented, else resort to
+        the method on the `initial_conditions` method on `CombinatorialClass.
+        """
+        logger.info("Computing initial conditions")
+        try:
+            return [
+                sum(
+                    Number(self.count_objects_of_size(n=n, **parameters))
+                    * reduce(mul, [var(k) ** val for k, val in parameters.items()], 1)
+                    for parameters in self.root.possible_parameters(n)
+                )
+                for n in range(check + 1)
+            ]
+        except NotImplementedError as e:
+            logger.info(
+                "Reverting to generating objects from root for initial "
+                "conditions due to:\nNotImplementedError: %s",
+                e,
+            )
+        return self.root.initial_conditions(check)
+
     def get_genf(self, check: int = 6) -> Expr:
         """
         Return the generating function for the root comb class.
@@ -252,21 +285,9 @@ class CombinatorialSpecification(
         """
         eqs = tuple(self.get_equations())
         root_func = self.get_function(self.root)
-        try:
-            logger.info("Computing initial conditions")
-            initial_conditions = [
-                self.count_objects_of_size(n=i) for i in range(check + 1)
-            ]
-        except NotImplementedError as e:
-            logger.info(
-                "Reverting to generating objects from root for initial "
-                "conditions due to:\nNotImplementedError: %s",
-                e,
-            )
-            initial_conditions = [
-                len(list(self.root.objects_of_size(i))) for i in range(check + 1)
-            ]
-        logger.info(maple_equations(root_func, initial_conditions, eqs,),)
+        logger.info("Computing initial conditions")
+        initial_conditions = self.get_initial_conditions(check)
+        logger.info(pretty_print_equations(root_func, initial_conditions, eqs))
         logger.info("Solving...")
         solutions = solve(
             eqs,
@@ -287,26 +308,24 @@ class CombinatorialSpecification(
                 return sympy.simplify(genf)
         raise IncorrectGeneratingFunctionError
 
-    def get_maple_equations(self, check: int = 6):
+    def get_maple_equations(self, check: int = 6) -> str:
+        """
+        Convert the systems of equations to version that can be copy pasted to maple.
+        """
         eqs = tuple(self.get_equations())
         root_func = self.get_function(self.root)
-        try:
-            logger.info("Computing initial conditions")
-            initial_conditions = [
-                self.count_objects_of_size(n=i) for i in range(check + 1)
-            ]
-        except NotImplementedError as e:
-            logger.info(
-                "Reverting to generating objects from root for initial "
-                "conditions due to:\nNotImplementedError: %s",
-                e,
-            )
-            initial_conditions = [
-                len(list(self.root.objects_of_size(i))) for i in range(check + 1)
-            ]
+        initial_conditions = self.get_initial_conditions(check)
         maple_eqs = maple_equations(root_func, initial_conditions, eqs)
-        logger.info(maple_eqs)
         return maple_eqs
+
+    def get_pretty_equations(self, check: int = 6) -> str:
+        """
+        Convert the systems of equations to a more readable format.
+        """
+        eqs = tuple(self.get_equations())
+        root_func = self.get_function(self.root)
+        initial_conditions = self.get_initial_conditions(check)
+        return pretty_print_equations(root_func, initial_conditions, eqs)
 
     def count_objects_of_size(self, n: int, **parameters) -> int:
         """
@@ -348,7 +367,11 @@ class CombinatorialSpecification(
         Raise an SanityCheckFailure error if it fails.
         """
         return all(
-            all(rule.sanity_check(n) for rule in self.rules_dict.values())
+            all(
+                rule.sanity_check(n, **parameters)
+                for rule in self.rules_dict.values()
+                for parameters in rule.comb_class.possible_parameters(n)
+            )
             for n in range(length + 1)
         )